pyproject.toml
[project]
name = "incident-response-env"
version = "0.1.0"
description = "Production Incident Response environment for OpenEnv — agentic RL training."
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "httpx>=0.24.0"]

[project.scripts]
server = "incident_response_env.server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["incident_response_env"]
